requirements.txt
pandas
numpy
pyarrow 
fastparquet 
duckdb 
dask 
scipy 
scikit-learn 
matplotlib 
seaborn 
plotly 
altair 
folium 
pydeck 
openpyxl 
pyjanitor 
dateparser 
unidecode 
requests 
httpx 
beautifulsoup4 
bs4 
lxml 
html5lib 
playwright 
tabula-py 
pdfplumber 
ijson 
tqdm 
s3fs 
fsspec 
patool 
Pillow 
deltalake 
aiofiles 
fastapi 
uvicorn 
python-multipart 
pydantic 
google-genai 
anthropic 
openai 
sqlalchemy 
networkx 
pymupdf
python-dotenv 
loguru 
orjson 
pytest 
docker
playwright>=1.40.0
python-Levenshtein
statsmodels